Renamed the dns_prefetch component to network_predictor
This is in preparation for adding preconnect support which goes through
the same underlying plumbing.  dns_prefetch was too restrictive of a
component name and was only exposing the single method.

There are no functionality changes in this CL, just the rename (the namespace was also renamed)

BUG=450682

Review URL: https://codereview.chromium.org/848303005

Cr-Original-Commit-Position: refs/heads/master@{#314001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed4b6f8b7392eb64fc664aa49f2627ecb5fdc073