Refactor the symbolize step of the orderfile generation.
This builds on the previous refactorings (linked in the bug), and unifies the
parsing of object files. It also removes the fuzzy matching of offsets done
that was previously also done in patch_orderfile.py.

It is also faster for three reasons:
- Elimination of an O(N^2) search
- Parallelization of object file parsing.
- No binary search

BUG=452879

Review URL: https://codereview.chromium.org/874683004

Cr-Original-Commit-Position: refs/heads/master@{#314137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 753ce83612800633fe2081c139df118ef94d7c41