telemetry: Fix timeouts for navigation.
The navigation methods in navigate.py and action_runner.py have a timeout, but
the timeout is only applied to the navigation, not to
WaitForDocumentReadyStateToBeInteractiveOrBetter. I modified both methods so
that the timeout applies to the entire method.

BUG=

Review URL: https://codereview.chromium.org/887993005

Cr-Original-Commit-Position: refs/heads/master@{#314475}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89c415666ec3492f1ae0851b9ea95c3a4cb9f38c