Adding UMA logging to touch text selection
For the Aura path we log the duration of the selection sequence -
from the the moment the text selection handles are brought up, and until
their dismissal, and also whether the selection was "successful" or not,
i.e. whether it ended with execution of one of the a command from the
quick menu.
For the Android path we only log the duration.

BUG=NONE

Review URL: https://codereview.chromium.org/895903003

Cr-Original-Commit-Position: refs/heads/master@{#315894}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e58cd2ca21f020e349893fbd310ade83cf27f47