Make SupervisedUserWhitelistInstaller work correctly with multiple clients (i.e. with multi-profile).
SupervisedUserWhitelistInstaller is now owned by the browser process and keeps
track of the union of all whitelists in all profiles. When a
SupervisedUserWhitelistService wants to register or unregister a whitelist, it
passes a client ID to the SupervisedUserWhitelistInstaller, so the latter knows
which clients (profiles) are interested in a given whitelist. The first time a
profile registers a whitelist with the SUWI, the SUWI registers the component
with the ComponentUpdaterService, and when the last profile unregisters the
whitelist, the component is unregistered from the CUS.

As client ID we use the last component of the profile path ("Profile 1",
"Profile 2", ...), which is also what ProfileInfoCache uses as a profile
identifier. It also lets us unregister all whitelists for a profile when it is
deleted, even if that profile has not been loaded into memory.

BUG=436459

Review URL: https://codereview.chromium.org/868363005

Cr-Original-Commit-Position: refs/heads/master@{#315969}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22951c3113400562410b454123bda3b2f05797af