[Telemtry] Move url lists of alexa1-10000 page to a json file.
Existing implementation of alexa1-10000 embeds the huge url list in
the python file, which is a terrible. It makes it super hard to navigate
through the file, or review its content, and make it harder to update the
alexa1-10000 pages if needed.

Review URL: https://codereview.chromium.org/915303002

Cr-Original-Commit-Position: refs/heads/master@{#316012}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fec4b0793f9d7eb6bef66f83687f10e6391dff62