Refactor Extension initialization logic. Add tracing and an additional latency histogram.
Note, this has a functional change on the Extensions.LoadAllTime metric, which didn't have the second half of the method (the part I put in LoadExtensionsMetrics()).
I expect that this section does not take very long, and it makes sense for the LoadAllTime metric to measure the entire method. If you're unhappy of the functional change of the metric, another approach is to add a histogram for LoadExtensionsMetrics() specifically, but I feel this is excessive.

BUG=454789

Review URL: https://codereview.chromium.org/959513002

Cr-Original-Commit-Position: refs/heads/master@{#318120}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8fd8d3b8b0356cf1fb480d90eb89eb8d8b36c4c