Adding telemetry test for page scroll using scrollbar drag.
This CL adds a telemetry benchmark to use gmail page and uses the
scrollbar and mouss drag action to scroll the page. The test calculates
the mid-point of scrollbar and drags it till the end of the page.

BUG=464685

Review URL: https://codereview.chromium.org/985603003

Cr-Original-Commit-Position: refs/heads/master@{#319629}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b874423d6e08083a2b78cb61f5992c27c92cc2d