IDL: Add support for serializer definitions
A serializer definition is output as a 'Serializer' production, with
either no children, an 'Operation' child, or a 'Map'/'List' child. The
production has an attribute named 'ATTRIBUTE' in the

  serializer = identifier;

form (and no children), and the 'Map'/'List' productions have attributes
named 'GETTER', 'INHERIT' and 'ATTRIBUTES', where the latter is a list of
the attributes named in the definition.

BUG=469650

Review URL: https://codereview.chromium.org/1031593003

Cr-Original-Commit-Position: refs/heads/master@{#321780}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 577bccbad01ca22a55df7d6c828459e39616d8fd