[Telemetry] Use new battery disabling method for telemetry.
A new way to enable collection of power data that is not
dependant on the divice was implemented in
https://codereview.chromium.org/993733002/
This CL is to change telemetry to use this when getting dumpsys data.
This method of disabling charging does not work with monsoon since
it does not actualy disable charging, but just starts the dumpsys collection.

BUG=

Review URL: https://codereview.chromium.org/1010213008

Cr-Original-Commit-Position: refs/heads/master@{#321945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f83d74abc7eb83445e074cec0925f57fb6f53dd2