[Extension API Extern Generation] Auto-generate enums
Enums used to simply be strings, but now we expose them (as string constants)
on the api object as part of the extension bindings. As such, it's appropriate
to have the externs generate an enum definition for them.

BUG=469920

Review URL: https://codereview.chromium.org/1015033003

Cr-Original-Commit-Position: refs/heads/master@{#322086}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d68755a4852b0aca4520df3a0ba7ceef76625d34