Telemetry] Do not close tab connections if the tab is already gone. Update unit tests so they check for formatted exceptions (so they will catch a bug such as this one).
This is the second attempt at a fix. The first attempt,
https://crrev/1041843002/, resulted in a flaky test (https://crbug.com/471512).

This fix is safer (skips CloseConnections call if the tab is not alive).

If this ends up having problems on Android, see if IsAlive is working properly. It has a bug on it https://crbug.com/463933

BUG=444240

Review URL: https://codereview.chromium.org/1048733003

Cr-Original-Commit-Position: refs/heads/master@{#322933}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb4c0abcbfe2989d36046bdc187f402501770e2c