[Telemetry] Rename interaction_record field of value to tir_label.
The name 'interaction_record' does not make it sufficiently clear that the
so-named field of Value is to be used to indicate the label of the
TimelineInteractionRecord to which it pertains. This CL renames it 'tir_label'
to make this distinction clear.

R=nednguyen
BUG=461101

Review URL: https://codereview.chromium.org/1057213005

Cr-Original-Commit-Position: refs/heads/master@{#323848}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f8bf3b33319a20fbe475f72c4a0d91d742c5252