ChannelMojo: Ensure that it always has ScopedIPCSupport
ChannelMojo has ScopedIPCSupport, but it is instantiated only in
in-process mode. This CL lets it always instantiate to make it
clear that ChannelInfo is protected by the ScopedIPCSupport.
It simplifies the relationship between the support object and
the channel, and makes the lifecycle invariant reasonable.

With this change, we no longer need to protect ChannelMojo with
ScopedIPCSupport on its client side. Now it's built-in.

Note that this is a speculative fix of fuzzer generated crash,
where Mojo channel related globals are gone before when channel
mojo is being destroyed.

BUG=473438
R=viettrungluu@chromium.org, rockot@chromium.org

Committed: https://crrev.com/808706e71c213c916815e967e3156cae64d12c00
Cr-Commit-Position: refs/heads/master@{#324308}

Review URL: https://codereview.chromium.org/1054253005

Cr-Original-Commit-Position: refs/heads/master@{#324488}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00194e78d7abbe83af8aab3e23c49e90d767126f