Fine-grained UMA for modules with out-of-order relocs.
The original code used 50 buckets for 100 count, so it wasn't possible
to know which index was being logged. This change switches to logging
the module index as if it were an enum, and labels the enum values for
friendlier analysis.

BUG=none
R=anthonyvd@chromium.org,asvitkine@chromium.org

Review URL: https://codereview.chromium.org/1083713002

Cr-Original-Commit-Position: refs/heads/master@{#325249}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0732d2d1fac6f20ccd7c0df5cdddd7a24e5576f4