Allow paths to be resolved in the GN args file
This uses the real settings object which contains the location of the
source root for resolving "//..." paths. The use-case is to allow
imports to be processed in the args file, which in turn allows one to assign
a set of build arguments by reference.

Review URL: https://codereview.chromium.org/1104313002

Cr-Original-Commit-Position: refs/heads/master@{#327131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c2f498dd8cb63c49663b5f80aabf808b4fd1d1c4