Measure the amount of time to open an extension popup starting, from when the user clicked.
There is already a metric "Extensions.PopupLoadTime2" which measures the amount
of time for the popup page to load. However, this doesn't take into account any
delay between clicking on a button and the popup opening, which may actually be
significant as we experiment with different queuing heuristics.

This CL adds a new metric "Extensions.PopupCreateTime" which measures the entire
time from clicking, not just the loading component. This needs to stay low.

BUG=477445
R=mpearson@chromium.org

Review URL: https://codereview.chromium.org/1125713006

Cr-Original-Commit-Position: refs/heads/master@{#328460}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f101acf0050b34a04e3de1d5c9274cd128637ab