Stop pulling in hunspell through DEPS (gclient sync).
Hunspell is now checked-in the Chromium tree (under
third_party/hunspell_new) and the dependencies are using it from there.
Thus we don't need to be pulling it from the external repo anymore.

After this lands, we are going to wait a week or so, to rename
third_party/hunspell_new back to third_party/hunspell to smooth the
transition and not get into the foot of the other devs.

BUG=472897
R=brettw@chromium.org
TBR=phajdan.jr@chromium.org # for checklicenses

Review URL: https://codereview.chromium.org/1130203003

Cr-Original-Commit-Position: refs/heads/master@{#329108}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fcbdb22486a7f84c56bb3b082029c7759648914